fix: fixed float inconsistency

+++ schemas/checkout.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import TypedDict
 
 
@@ -5,13 +6,13 @@ class CheckoutItemData(TypedDict):
     productId: int
     name: str
     quantity: int
-    unitPrice: float
-    lineTotal: float
+    unitPrice: Decimal
+    lineTotal: Decimal
 
 
 class CheckoutContextData(TypedDict):
     sessionToken: str
     tableId: int
     items: list[CheckoutItemData]
-    subtotal: float
-    total: float
+    subtotal: Decimal
+    total: Decimal
