+++ schemas/order.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import TypedDict
 
 
@@ -6,5 +7,5 @@ class OrderItemInsert(TypedDict):
     product_id: int
     name: str
     quantity: int
-    unit_price: float
-    line_total: float
+    unit_price: Decimal
+    line_total: Decimal
